Adding initial version of the scenario refactoring
  * Read scenario code refactored

  * Write scenario pending to be refactored

  * Scenario tests need some refactoring too
