+++ benchmark/scenarios/__init__.py
@@ -1,8 +1,5 @@
 from .no_load import NoLoadScenario
-from .read_request_load import (
-    ReadRequestLoadScenario, RequestRateTooLow, RequestRateNotReached,
-    RequestOverload
-)
+from .read_request_load import ReadRequestLoadScenario
 from .write_request_load import (
     WriteRequestLoadScenario, WRequestRateTooLow, WRequestRateNotReached,
     WRequestOverload, DatasetCreationTimeout
@@ -10,3 +7,8 @@ from .write_request_load import (
 from .rate_measurer import (
     RateMeasurer, DEFAULT_SAMPLE_SIZE
 )
+from .request_load import (
+    RequestLoadScenario, RequestRateTooLow, RequestRateNotReached,
+    RequestOverload
+)
+
